requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
openai-whisper==20250625
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.22
regex==2026.2.28
requests==2.32.5
setuptools==82.0.1
starlette==0.52.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
